Data/Vercel/requirements.txt
fastapi
httpx
markdownify
numpy
semantic_text_splitter
argparse
tqdm
uvicorn
google-genai
google.generativeai 
openai
pydantic
pillow